[FIX] l10n_de_pos_res_cert: prevent divide by zero
**Current behavior:**
Adding a product to a POS order then zero-ing out the quantity
of the line will result in an instance of division with this qty
(now zero) in the denominator.

**Expected behavior:**
Don't allow a scenario to occur where 0 can appear in the
denominator of a division operation.

**Steps to reproduce:**
1. Switch to the demo German company installed with
     `l10n_de_pos_res_cert`

2. Enable Fiskaly in the company settings for the German
     company

3. Add a payment method in the POS config

3. Add a POS of type Restaurant and enable the TSS toggle
     setting (comment out the  `_l10n_de_fiskaly_iap_rpc()` call
     in `l10n_de_pos_cert/models/pos_config.py` to save)

4. Add a floor and a table to the restaurant

5. Start a POS session and start an order, add some product
     then select the order line and zero out the quantity with
     the digital numpad

6. Go to payment -> register payment -> validate

7. Observe the error dialog

**Cause of the issue:**
In `l10n_de_pos_res_cert/models/pos_order.py`, the
`remove_from_ui` method calls `_line_differences()` if fiskaly
is setup for our German company which makes the call to
`_merge_order_lines()` where the following operation:
`str(group[0]['price_subtotal_incl']/group[0]['qty'])` may have
a denominator equal to 0.

**Fix:**
Do not call `_merge_order_lines()` on order lines which have a
quantity of zero.

opw-3751492

closes odoo/enterprise#60795

X-original-commit: bccfb6fff199e3ac264d76e5e9bbf6707d88b84c
Signed-off-by: David Monnom (moda) <moda@odoo.com>
Signed-off-by: Vincent Ethan <etvi@odoo.com>