[FIX] helpdesk_timesheet: ensure the timesheet date is the current one
Before this commit, the timesheets created inside `test_timesheet_check_warning_when_helpdesk_team_change`
test could be unvalidated even after calling `action_timesheet_validate`.
The reason is because the timesheets are created by the current user
(OdooBot) is our case and his timezone could be different than the user
who validates the timesheets and so the timesheets date could be
tomorrow for the timesheet approver used inside that test becaue of the
timezone is not the same for the both users.

To avoid timezone issue, this commit makes sure the same user creates and
validates those timesheets since the goal of that test is to check the
`project_id` field is unchanged for validated timesheets even if the
ticket changed.

runbot-61011

closes odoo/enterprise#60264

X-original-commit: 86996486b6cf83fa7e14f4c79ab2171965ea691f
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>