[FIX] helpdesk: fix ticket response grouping issue
Some MUA allow users to group messages 'by thread', so that all emails
that share some specific headers are bundled together in a
'conversation' view.

In general, using the 'References' header to point to the original
message in the thread is enough for most MUAs to properly group emails
together.

This is not the case for Outlook, which also wants the title to somewhat
match the original title - some changes are accepted (typically at the
start of the subject header) but others aren't.

For example, if a ticket is submitted by email alias with the title
'Important issue', then replies with titles like 'Re: Important issue'
or '(#123) Important issue' are recognized as being from the same
thread, but others like 'Important issue (#123)' are not.

Since the helpdesk override of the display name is mostly there for the
backend and not for email communications, this commit forces the
helpdesk app to send emails with the exact same title as the initial
ticket instead of including the ticket reference in the reply's title.

The ticket reference is still included in the mail body for the customer
so no information is lost.

opw-3748509

closes odoo/enterprise#60202

X-original-commit: 0fd53102589e2bdc4d4a257f367a37eca88604c8
Signed-off-by: Bouvy Damien (dbo) <dbo@odoo.com>