[FIX] website_enterprise: fix web_enterprise dependency
Since [1], when attempting to uninstall web_enterprise, the webclient
will no longer start with website_enterprise installed. The reason for
this is because the patch to color_scheme_service imports the service
from web_enterprise, but __manifest__ was not updated to mention the
dependency and therefore, the file was no longer loaded by the
framework.

On top of being the expected fix for new databases, the databases whcih
are broken due to this issue will be indirectly fixed too by just
restarting the server after pulling this fix. No module update or
anything else will be required.
Indeed, the odoo framework won't try to load installed addons if one of
their dependencies is not installed / missing. In this case, it means
that `website_enterprise` won't be loaded and the problematic JS file
won't be a problem.

Usually, this kind of changes are done in master and the users are told
to re-install the app, but in this case, they cannot log into the
webclient to do so. Therefore, the decision is made to do this change in
17.0, where the faulty line was introduced.

[1]: https://github.com/odoo/enterprise/commit/90df39603b07

closes odoo/enterprise#60231

Signed-off-by: Romain Derie (rde) <rde@odoo.com>