[FIX] sale_purchase_inter_company_rules: apply putaway strategy on receive
Steps to reproduce:

With company 1:

- Go to the General Settings and activate:
  1) Inter-Company Transactions:
     - "Synchronize Sales and Purchase Order"
     - "Automatic Validation"
     - "Copy Lots on Delivery Validation"
  2) Multi-Step Routes
- Go to Inventory > Configuration > Warehouse Management > Locations
- Create a new Location SHELF with COMP1/STOCK as parent
- Go to Inventory > Configuration > Warehouse Management > Putaway Rules
- Create a new putaway rule applicable to all products from
  COMP2/STOCK to COMP2/STOCK/SHELF

With company 2:

- Create a storable product tracked by a Unique Serial Number
- Click on "On Hand" > Register 1 on hand qty with a new serial Number
- Create and confirm an SO for company 2 with your product

At this point, if you go to company 1:

A purchase order related to the SO was created and confirmed. If you
open the related delivery and inspect the related stock move line the
destination, it will correctly be set to COMP1/STOCK/SHELF.

With company 2:

- Open the delivery linked to your SO click on the list icon of you
  stock move line
- Add a stock picking referencing the serial number you created
- Validate the delivery

If you go to company 1, the move line is now attached with a serial
number but the destination is set to set to COMP1/STOCK.

> The putaway rule is not applied anymore

Cause of the issue:

When you confirmed the SO with company 2, it will create an PO in
company 1:
https://github.com/odoo/enterprise/blob/3190d36192def6c63c254e16e754b12353a865e6/sale_purchase_inter_company_rules/models/sale_order.py#L22
and a related stock move will then be created. During the
"_action_confirm" of this stock move the putaway rules will be applied
to the stock move lines a few lines after creation:
https://github.com/odoo/odoo/blob/58fefc94ca39f7951392063921fe9b5e606178e6/addons/stock/models/stock_move.py#L1720
When you confirm the delivery, this correclty set stock move line will
be deleted and a new one will be created:
https://github.com/odoo/enterprise/blob/3190d36192def6c63c254e16e754b12353a865e6/sale_purchase_inter_company_rules/models/stock_picking.py#L30-L33
However, as the "apply_putaway_strategy" method is not called after the
creation of this newly created stock move line, the destination of the
line will entirly be set by the destination of the stock move:
https://github.com/odoo/odoo/blob/42b760c8fa27cfc75fcb2fd5b232defc9002e85f/addons/stock/models/stock_move.py#L1441
so that the put away rules will not be applied anymore.

opw-3744375

closes odoo/enterprise#62851

X-original-commit: 5f8825728922636531dfdd3ed7ba56bade328fcb
Signed-off-by: William Henrotin (whe) <whe@odoo.com>
Signed-off-by: Lancelot Semal (lase) <lase@odoo.com>