[IMP] account_online_synchronization: Show banner when fail connecting to existing account
Issue: Sometimes the connection with the bank breaks for some reasons either the user remove the connection from the bank dashboard
etc. And when the user press on 'fetch_transaction' in odoo, We will ask him to add the bank again without showing him anything
to indicate the reason.

Solution: Adding a banner in the widget to report that we are facing an issue connecting to the bank-account and show it
whenever the user face a problem connecting to existing account through the '_handle_odoofin_redirect_exception'
which handle open the iframe whenever we face an issue connecting to an existing bank account

Task-3355270

closes odoo/enterprise#69009

X-original-commit: 339e17c5123faf12587e45639db33dd5933991ef
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>
Signed-off-by: Muhamed Abdalla (muab) <muab@odoo.com>