[FIX] appointment: style not correct on small event (15 min)
For info FullCalendar V6 use Preact internally.
Preact is similar to OWL/React/Vue as when some props change it apply
the change on the element linked to the props.
In our case for the events, when an event has a duration of 15 minutes
FullCalendar changes the `isShort` to `true` so Preact can add the
`fc-timegrid-event-short` class to the event.

In Odoo we add also other classes to these events for our needs
(e.g.: `o_event_striked`).
In FullCalendar V4 it was done in `eventRender` using
`el.classList.add()`. During the migration to FullCalendar V6 [1]
the `eventRender` was changed into `eventDidMount` (a Preact HOOK).

In most cases, `eventDidMount` is fine but not for short event, as
in FullCalendar V6 there is another HOOK especially for the classes
`eventClassNames` [2].

Before this fix, the render flow of the event was:
1) pre-render event
2) add the event to the DOM
3) call `eventDidMount` (here we add our classes)
4) FullCalendar set `true` to `isShort` props
5) Preact sets the classes on the event to add `fc-timegrid-event-short`
   (here FullCalendar removes our classes added in `eventDidMount`)

After the fix, the render flow is the same as before, but as we add the
classes using the `eventClassNames` HOOK, FullCalendar knows all
additional classes per event so when Preact (in steps 5) sets the
classes its adds `fc-timegrid-event-short` and all other classes added
by our code in `eventClassNames`.

Steps to reproduce:
* Open Calendar App
* Make an event with a duration of 15 minutes
* Refresh the page (F5)
  => Bug the style of new event (15 min) is wrong

[1]: odoo/odoo@90f85a19deaea33cd747c969762ff20f1d59ef4c
[2]: https://fullcalendar.io/docs/event-render-hooks

closes odoo/enterprise#60293

X-original-commit: ed4ae194e7bea4bdadf1b0e053fa040cd705f887
Related: odoo/odoo#161019
Signed-off-by: Pierre Paridans (app) <app@odoo.com>
Signed-off-by: Romeo Fragomeli (rfr) <rfr@odoo.com>