[FIX] web_gantt: repair resize in RTL
**Before this commit**
- Apply right-to-left direction in the language in use
- Go to a gantt view with resize enabled (i.e. Planning)
- Resize do not work

**After this commit**
The resize is working.

**Notes**
Previously, the applied style to the resized pill and the resize badge
were computed based on the "isStart" handle.
The problem was that "isStart" is dependant of the UI direction.
In LTR, "isStart" means the left handle and the right one in RTL.
This commit always takes the left or right handle into account,
regardless of the UI direction.

No test has been written as we currently have no easy way to mock a
proper RTL interface. For now, we do not have a test suite that has
assets bundles processed through rtlcss.

closes odoo/enterprise#60344

X-original-commit: bb170b878e73f3134e6de08e7395c25fa2e60986
Signed-off-by: Julien Mougenot (jum) <jum@odoo.com>
Signed-off-by: Bruno Boi (boi) <boi@odoo.com>