[IMP] documents_spreadsheet: detect spreadsheet json files
In debug mode, you can download the spreadsheet json file from the
spreadsheet itself with the menu File > Download as JSON

In the Documents app however, try uploading the json file:
=> it's not recognized as a spreadsheet document.

Advanced users such as our own business analysts or ourselves
often download the spreadsheet files and re-upload it.

One use case is to build a spreadsheet on a test/staging
database, then upload the spreadsheet to the production server.

Current limitation (to keep it simple): uploading such a file
on an existing document doesn't work.

closes odoo/enterprise#60274

Task: 3596641
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>