[FIX] whatsapp_event: Include message to unregistered attendee in total sent count
Before this commit the Whatsapp Sent Count (located in the Communication
tab of the event view) wasn't including messages sent to the attendees who
weren't confirmed.

[This commit changes]
Event mail scheduler will include seats_uncofirmed into calculation of
the mail_count_done which is the count of emails sent. This is
reflecting the actual number of messages sent.

[Reproduce]
- Install whatsapp_event
- Create event E starting in time T (Events/Events)
- Add an "Unconfirmed" attendee
- Add an E-Mail in the Communication tab of the event E
	- Set Interval time > T,
	- Template: "Email Reminder"
	- Trigger "Before the event"
- Run scheduled Action sending whatsapp (Scheduled Actions / Event: Mail Scheduler)
- BUG: message sent, but not included in the sent count in the communication tab of the event E

Related to the commit in the odoo community called:
[FIX] event: Include unregistered attendee emails in total sent count

opw-3693626

closes odoo/enterprise#60261

X-original-commit: 23c36bb6c9907004658820e0e008d2e1ec31918a
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>
Signed-off-by: Andrzej Pietrusiak (pian) <pian@odoo.com>