[FIX] web_studio: add action button with xml_id
In a form view, add a stat button in the button box.

Before this commit, the button contained the action's id. It worked on a single DB
but when exporting, the id might have changed.

After this commit, we put the xml_id of the action instead, which is set when studio=1 is in the context.

A similar fix is done for action buttons in form and list header, feature that has been added at 025829afaacb8e02bae10e67de7dca0f52fc5d30

opw-3824053

closes odoo/enterprise#60789

X-original-commit: eebb1a6f8d3ec076d34faf7e9e820e257d5be5db
Signed-off-by: Bruno Boi (boi) <boi@odoo.com>
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>