[FIX] pos*: fix pos_self_order_preparation_display has_paper field
pos*: pos_preparation_display, pos_self_order_preparation_display

A previous commit has introduced a field `has_paper` in the pos_config.
This field is introduced in the module pos_self_order and should only be
used in this context. The problem here is that it was used in the
pos_preparation_display module which is not correct. This commit fixes
this by moving the changes brought by the first commit in the correct
module 'pos_self_order_preparation_display'.

closes odoo/enterprise#60395

Signed-off-by: Vlad Stroia (vlst) <vlst@odoo.com>