[FIX] planning: ensure the filter stay to highlight the next action
Before this commit, when the user clicks on `Copy Previous` in the gantt
view of `planning.slot`, a filter will be added to highlight the shifts
added by the action executed. However, if the user clicks on `Auto Plan`
to plan the open shifts automatically, the filter is disappeared and no
shifts are highlighted as before. The reason is because we toggle the
filter without checking if that filter is already there in the search
view.

This commit checks if the filter is already there before toggle it. If
yes, then we notify the model instead of toggling the facet in the
search view.

runbot-55489
runbot-58209
runbot-56675
runbot-58149

closes odoo/enterprise#60585

X-original-commit: 8f941ddcd75b34b2b7d2d1ab10b1e1f67aa3ac3f
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>