[FIX] planning: fix issue when creating shifts with multiple company records
Versions:
-master

Steps to reproduce:
-Go to the planning module
-Activate more than one company
-create Sales order in company A
-create shift in company B and select sales oder of comapny A
-save the record
-traceback

Issue:
On saving, when more than one company is activated and a shift is created for a
specific company with an record SOL selected from other company, a Singleton
Error occurs.

Cause:
The traceback arises due to inability to iterate over multiple companies
during this process on creating a shift.

Solution:
Implemented a looping to iterate over each company, ensuring proper handling of
the multiple company on the shift

task-3859050

closes odoo/enterprise#61583

X-original-commit: a8babdac6148d8b225619284fe33b5268d114afa
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>