[FIX] mrp_workorder: tours build errors
Before this commit, the test tour `test_generate_serials_in_shopfloor`
failed when runned without the demo data.
The issue was the admin user was not part of the
`mrp.group_mrp_routings` group so workcenters were displayed.

Before this commit, tour `test_shop_floor` was failing for the same
reason but would also fail because it uses an employee created in demo
data. Some employees are now created in the test to avoid that.

Runbot build errors 59833 and 60709

closes odoo/enterprise#60427

X-original-commit: c2748aa6d8b59808741f00dce9560be8e69eba64
Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>
Signed-off-by: Steve Van Essche <svs@odoo.com>