[FIX] pos_iot: Prevent TypeError while uninstalling IoT for PoS
Currently, an exception is generated when the user tries to uninstall
IoT for PoS.

Error:
`TypeError: uninstall_hook() missing 1 required positional argument: 'registry'`

This is because the commit [1] added the uninstall hook at 15, and the uninstall
hook requires two arguments in 15.0. But from saas-16.3 uninstall hook require
only one argument as 'env'; it is not changed with commit [1]'s forwarded port.

This commit will fix this issue by providing the argumnet 'env' that is
required in the uninstall hook.

[1]-https://github.com/odoo/enterprise/commit/6b10cc80ea2441b5b2ab86aab52abbf7084d4319

sentry-5167509820

closes odoo/enterprise#60989

X-original-commit: 7f61a9331627f31e2604478fbd2996034766017d
Signed-off-by: Robin Engels (roen) <roen@odoo.com>
Signed-off-by: ANSARI MAHAMADASIF (maan) <maan@odoo.com>