[IMP] knowledge: increase article width on insertion of a wide table
On computers (or any device capable of seeing the TablePicker), tables
inserted with the /table command now trigger full article width if they
have at least 5 columns.

Note that this cannot happen with tables inserted on mobile devices,
as they have a standard fixed size of 3x3.
Existing tables onto which more columns are added will not modify
article width.

Tables with a lot of columns are hard to read with the default article width,
and they did not previously increase that width on insertion.

task 3847398

Part-of: odoo/enterprise#60397