[FIX] sign: simplify completed doc gen./download
**Current behavior:**
When downloading sign documents via buttons in the form and tree
views, user will get an access error on the read perm for the
`sign.request.item.value` model.

**Expected behavior:**
These buttons should permit the download of the corresponding
document.

**Steps to reproduce:**
1. In the Sign app go to Documents

2. In the tree view click the download button by a document
     which has not already been generated (e.g., any of the demo
     data documents)

**Cause of the issue:**
The download buttons call `get_completed_document()` which, if
the completed doc has not been generated, will directly call
`_generate_completed_document()` where we do a `read_group()` on
the `sign.request.item.value` model. This model must be read
with `.sudo()` because access rights are never granted to any
group.

**Fix:**
Get rid of the `_generate_completed_document()` call in
`get_completed_document()`, letting the program go to the
`act_url` flow, into `download_document()`. Here we will access
the `sign.request` record with `env.su is True` which will
permit the `read_group()` later on in the aforementioned flow.
Because we always route from `get_completed_document()` with
a `download_type='completed'`, there is no reason to generate
the document here.

In the case of downloading many records in one action, we will
call `_generate_completed_document()` with sudo, when needed, in
the respective route, in the controller for this module.

opw-3834177

closes odoo/enterprise#61869

X-original-commit: 45afb2de74237f52daf08c7e14ea5c1c630ad4f3
Signed-off-by: Vincent Ethan <etvi@odoo.com>