[FIX] documents: fix inspector scrollbar
Purpose
=======
Fix the unnecessary horizontal and vertical scrollbars appearing in the
inspector when a document name is too long.

Specifications
==============
In the inspector, when the value of the fields of type model (like the Journal Entry
or Signature fields) is too long, the value is styled so that its width takes the max-width.
However the field max-width is incorrectly computed. It is missing the table-sm class
horizontal padding. The field is thus too big to fit in the inspector and scrollbars appear.

Taking into account the table-sm paddings while computing the field value max-width.
As the field label left padding is set to 0, the table-sm padding is multiplied by 3
(right padding of label, left and right padding of value).
Also adding the "overflow-hidden" class on the inspector so that scrollbars never appear.

Task-3829716

closes odoo/enterprise#60501

X-original-commit: 876c207b15d8a1e84eecc66c01a1bc8588e19ab4
Signed-off-by: Stéphane Debauche (std) <std@odoo.com>
Signed-off-by: Amélie Dieudonné (amdi) <amdi@odoo.com>