[FIX] sale_timesheet_enterprise: fix search_read on hr.employee
Currently, in the timesheet leaderboard timer grid, kanban and list
views, a search_read is done on hr.employee to retrieve the
billable_time_target field. However, this field is exclusive to users
that have the hr.group_hr_user, and so if a user without that group
tries to open any of these views, a traceback is raised.

To fix this, a method was added to retrieve the billable time targets
using a sudo().

closes odoo/enterprise#60463

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>