[FIX] sale_subscription: respect maximum_amount when paying subscription
Previously, it was easy to bypass the maximum_amount field on
payment.provider by going to a subscription and clicking "Set Payment
Method". The subscription would then be charged with the token,
regardless of the maximum_amount set on the payment.provider.

This filters out ineligible acquirers through
_get_compatible_providers() so it's not possible to use them for new
payment methods. The existing override wasn't working because
sale_order_id was a string and sale.order('1').exists() always returns
False. Casting sale_order_id to a string was added in a small
/my/payment_method route override.

Filtering out already existing payment methods is harder because the
sale order isn't available in _get_available_tokens(). Instead we
raise a UserError when a user assigns an ineligible payment method.
This is handled well in the frontend.

This focuses only on customer-facing flows. The maximum_amount field
isn't checked either on e.g. account.payment but that's only
accessible to internal users. Other customer-facing flows aren't
affected because typically _get_available_tokens() is called with
eligible payment providers as its first parameter.

Attempted alternative approaches:
- odoo/odoo#161021, odoo/enterprise#60294
- odoo/enterprise#60382

opw-3858340

closes odoo/enterprise#60930

X-original-commit: 0a7eaf4d9c21768530d7ce12232b47401f9f9952
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>
Signed-off-by: Joren Van Onder (jov) <jov@odoo.com>