[IMP] account_followup: Added Pay now button in followup report
A "Pay now" button for overdue invoices was added to the "Payment Reminder" email template.
The button is only visible if "invoice online payment" was activated in the settings, and if there is some available payment method.

Before this commit, the follow-up report was used to remind customers they have unpaid invoices,
but they still had to pay invoices manually which is not very friendly and does not push them to pay all of them.

task-3386024

closes odoo/enterprise#61201

Related: odoo/odoo#162762
Signed-off-by: Quentin De Paoli (qdp) <qdp@odoo.com>