[FIX] l10n_mx_edi: avoid using the demo data company in tests
When importing a CFDI, we try to retrieve the partner (see
`_retrieve_partner`). If more than one partner is retrieved, the
partners are not returned.

During the `_l10n_mx_edi_import_cfdi_fill_partner` test, there exists
two partners with the same VAT (one has a company_id, the other
doesn't). Hence, when importing the bill, the `_retrieve_partner` will
find 2 and will not return anything, and we end up creating a new
partner.

To fix that, we remove the VAT of the partner from the demo company.
Thus, only one partner will be retrieved.

opw-3829223

closes odoo/enterprise#60928

X-original-commit: 40be8f77afc0aca84ccf7e9062e27ebcda1aad1e
Related: odoo/odoo#162208
Signed-off-by: William André (wan) <wan@odoo.com>
Signed-off-by: Julien Van Roy (juvr) <juvr@odoo.com>