[FIX] web_studio: reportEditor: avoid bad query crashes when guessing variables
To be able to render a human-readable qweb, we need to evaluate some arbitrary python
expressions. Thos can produce bad sql queries because at this point, we are not, by construction
in the conditions that produce correct queries (missing records etc....)

Before this commit, when an query error occured at that step, the whole transaction was corrupted, and crashed
the whole thing.

After this commit, we do our partial evaluation things in a separate cursor, to avoid polluting the main one.

opw-3815284

closes odoo/enterprise#61795

X-original-commit: 0eac0551501a32977bd3b3f70b46d186044b0eba
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>