[IMP] account_3way_match: better xpath specificity
Currently, an inherited view for `account.view_move_form` targets field
`fiscal_position_id` directly with an xpath.

While that works for standard modules, it is a very broad xpath that
only succeeds in selecting the correct field because it appears before
other ocurrences of identically named fields.

This commit narrows down the xpath to make it explicit that we target
the field located in `account.move`'s "Other Info" page.

This fixes an ongoing issue with a customization (described in the task)
while improving readability and extensibility of
`account_invoice_form_inherit`.

Task-3837027

closes odoo/enterprise#60462

X-original-commit: 55e1661ac2c9342092124fc0481eb0909365f4a8
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Stanislas Gueniffey (stgu) <stgu@odoo.com>