[IMP] account_3way_match: better xpath specificity
Currently, an inherited view for `account.view_move_form` targets field
`fiscal_position_id` directly with an xpath.

While that works for standard modules, it is a very broad xpath that
only succeeds in selecting the correct field because it appears before
other ocurrences of identically named fields.

This commit narrows down the xpath to make it explicit that we target
the field located in `account.move`'s "Other Info" page.

This fixes an ongoing issue with a customization (described in the task)
while improving readability and extensibility of
`account_invoice_form_inherit`.

Task-3837027

closes odoo/enterprise#60979

X-original-commit: 4d38831f6c598e64e82d0ae6b9c9114f4461efe8
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Stanislas Gueniffey (stgu) <stgu@odoo.com>