[FIX] pos_preparation_display: prevent splited order display
Prior to this commit, splitting an order would create a new order,
causing it to appear again on the preparation display. This could
lead to the kitchen preparing the same order twice. The sequence of
events was as follows:

1. The order is placed.
2. The kitchen receives and prepares the order.
3. The waiter delivers the order to the table.
4. The client receives the bill and requests a split.

At the point of splitting, a new order is created. This duplicate order
should not be sent to the kitchen as it represents a meal that has
already been prepared and consumed. This commit resolves this issue by
preventing display of duplicate orders to the kitchen during order
splitting.

Community PR: https://github.com/odoo/odoo/pull/161498

opw-3809693

closes odoo/enterprise#60537

Related: odoo/odoo#161498
Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>