[IMP] appointment: add invert interval util
inverting intervals is a nice way to get "unavailabilities" from a list of
"availabilities" which can be nicer to work with for some use cases.

The goal here is to get intervals where appointment slots are not defined
from a list of intervals where appointment slots are defined.

tests for both this and `interval_overlaps` are added

task-3813567

Part-of: odoo/enterprise#60100