[IMP] appointment: display slot unavailabilities
User slots are not reflected in the gantt view so there
are no indications of whether an appointment is within a valid
time period with regards to the appointment-defined time intervals.

This commit adds slot intervals in gantt availabilities display and
attempts to have more common ground between slots and gantt.

task-3813567

Part-of: odoo/enterprise#60100