[IMP] account_reports: introduce custom values in groupby
Before this commit, the values allowed into the groupby field of account.report.line were restricted to only fields of account.move.line. This, however, was a bit too restrictive for some custom reports we'd like to convert to custom engines (in our case, Instrastat Report), so we decided to solve that by introducing a new hook on the custom handlers, allowing the definition new custom values for the groupby, to use them later on in some custom engines.

closes odoo/enterprise#60565

Signed-off-by: John Laterre (jol) <jol@odoo.com>