[FIX] pos_preparation_display: fix pos config and prep display conn
Issue -->

When creating a POS order via any pos config, any active preparation display that has any of the product categories used in the order will cause the customer display to reload. This happens regardless of the `pos_config_ids` set on the preparation display record. Therefore, an order from a `pos.config` **not** on the preparation display will still cause the display to reload

Solution -->

Edit the search domain on `pos_preparation_display.display` to check the origin of the order to keep the display from reloading unnecessarily.

opw-3838879

closes odoo/enterprise#60664

X-original-commit: 16ba33cd77413948308dd09326021a14f52f51f6
Signed-off-by: Vlad Stroia (vlst) <vlst@odoo.com>
Signed-off-by: Aditya Dasgupta (adda) <adda@odoo.com>