[FIX] l10n_ar_edi: remove unit test for bono fiscal
It is not working because some issues with AFIP test envrionment
that added new restrictions that need to adapt-

In this case we depend on AFIP to solve it and we do not know when
we will have a solution. After everything is ok we will re add
the test.

closes odoo/enterprise#60864

X-original-commit: 6d250912835c85b2a4d1b094f1e3a4fab0d1738e
Signed-off-by: Josse Colpaert <jco@odoo.com>