[FIX] documents: fix tags count update
Purpose
=======
Fix the tags count which isn't updated in the search panel when:
- adding/removing the tag from a document using the inspector
- dropping a documents on a specific tag in the search panel

Specifications
==============
When adding/removing a tag to/from a document using the inspector or
dropping a document on a tag in the search panel, the record tag
was updated but the search panel wasn't notified of the tag update.
Reloading the search model after the tag update to modify the tag count.

Technical point:
Reloading the search panel instead of notifying it as a reload
is achieving the same result with less rpcs calls.

Task-3837247

closes odoo/enterprise#60615

X-original-commit: 8082b8ba509861256235f06b3b6636f7c54a1201
Signed-off-by: Stéphane Debauche (std) <std@odoo.com>
Signed-off-by: Amélie Dieudonné (amdi) <amdi@odoo.com>