[FIX] stock_barcode: bigger control button
Before this commit, the footer buttons in the Barcode app are too small.
This commit undoes that, making the buttons' size more or less the same
whatever the device's size or orientation.
That said, their height are still smaller in case the device height is
smaller than 800 px.

closes odoo/enterprise#60881

Task: fp request
X-original-commit: fc73fb93833a15f88a7c68155080171c778fcc1b
Related: odoo/odoo#162125
Signed-off-by: Quentin Wolfs (quwo) <quwo@odoo.com>
Signed-off-by: Steve Van Essche <svs@odoo.com>