[FIX] sign: sign template edition popover not closing
Before this commit:
when editing a sign template and opening the sign item popover,
it remained open when clicking away even when leaving the sign app.

This commit addresses this issue by refining the `closeOnClickAway`
to close the popover only if the click is not within an active modal,
ensuring that the popover stays open when the role modal is displayed.

closes odoo/enterprise#61724

Task: 3861496
X-original-commit: a2db6a75d6b45d7e204f202e7f32c4227394a03c
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>
Signed-off-by: Ahmad Almaghraby (alah) <alah@odoo.com>