[FIX] stock_barcode: inventory line buttons
Before this commit, there was an issue with the line's buttons for the
Inventory Adjustment in the Barcode app.
The -1/+1 buttons were placed above/below each other instead of being
on the same line, causing the edit button and the set button to be
misplaced on the left.
This commit fixes that, now these buttons will be placed correctly:
- On the same height if there is only two buttons;
- On two lines if there is four buttons.

closes odoo/enterprise#61077

X-original-commit: 0d2e011626b8877d231939ba09cbd3d88d089ca6
Signed-off-by: Quentin Wolfs (quwo) <quwo@odoo.com>
Signed-off-by: Steve Van Essche <svs@odoo.com>