[IMP] l10n_be_hr_contract_salary: require insurance notes
With this commit, hospital and ambulatory insurance notes are now
required when (and only when) the respective insurance options are
selected AND either a spouse or a relative child/adult is chosen as
extra beneficiaries for this insurance. In this case the notes are
needed to provide the identities of this spouse and/or other relatives.

This is not implemented in a generic way, as it is only needed for
Belgian payroll and salary benefits (as opposed to personal information)
usually have sensible default values. In addition, the insurance fields
are currently already handled with custom javascript.

task-3866298

closes odoo/enterprise#60630

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>