[FIX] account_consolidation: consolidation report: properly print the pdf in landscape mode when the report is too large
Before that, the report was always printed in portrait mode. This was due to the fact the consolidation report does not use the options['columns'] key, unlike pretty much every other report. Because of that, the standard code couldn't detect it would be too large for the portrait mode. We fix that by introducing a context key to control that, and forcing it when necessary on the consolidation report.

OPW 3850656

closes odoo/enterprise#61160

X-original-commit: ce2272eef9bdc35625b1f2417c7429d201dcb1d9
Signed-off-by: Laurent Smet (las) <las@odoo.com>
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>