[FIX] account_consolidation: consolidation report: properly print the pdf in landscape mode when the report is too large
Before that, the report was always printed in portrait mode. This was due to the fact the consolidation report does not use the options['columns'] key, unlike pretty much every other report. Because of that, the standard code couldn't detect it would be too large for the portrait mode. We fix that by introducing a context key to control that, and forcing it when necessary on the consolidation report.

OPW 3850656

closes odoo/enterprise#61195

X-original-commit: 4367042c8c7f2190d23019e591c332045b7d4993
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Laurent Smet (las) <las@odoo.com>