[FIX] helpdesk: fix ticket description
To reproduce:
=============
- configure helpdesk team with email address
- contact this email address with a contact doesn't exist in the database
-> the created ticket won't have a description

Problem:
========
the message's `author_id` is not set, which prevents the description to be set,
as we compare the `partner_id` of the ticket with the message's `author_id`
to make sure to deal with real customer and not Odoobot or anything else.

Solution:
=========
instead of comparing the `partner_id` of the ticket with the message's `author_id`,
we compare the `partner_email` of the ticket with the message's `email_from`.

opw-3850228

closes odoo/enterprise#61591

X-original-commit: d9590ab05b989460393473eb0815a02d58781d84
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: Abdelouahab Laaroussi (abla) <abla@odoo.com>