[FIX] account_reports: posting tax closing entry on composite tax report
Steps to reproduce
------------------
* install `l10n_lu`
* open the Tax Report
* create a tax closing and attempt to post it

You will be met with a traceback: `Trying to dispatch an action on a
report not compatible with the provided options`

Cause
-----
The bug occurs when attempting to post a tax closing for a tax report
that includes sections. This issue originates from the behavior of the
`get_options()` method, which, when invoked on a report containing
sections, retrieves the `report_id` of a section instead of the main
report's ID. This discrepancy occurs because of a feature in the code
known as 'rerouting', which intentionally changes the `report_id` to that
of the section.

Consequently, when an action is dispatched to the main report using
these options, an error is triggered due to the mismatch between the
expected report ID and the one provided through the options.

Fix
---
To resolve this issue, the fix involves utilizing the
`on_sections_source` parameter when dispatching the action for tax
closing. By setting this parameter to true for composite reports, it
effectively disables the rerouting behavior, ensuring that the action is
directed to the correct report ID, thus avoiding the error.

opw-3816175

closes odoo/enterprise#60697

X-original-commit: 48481cc2063b692a4d68d5f7d84842e0bcb6ac65
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Séna Serge Nshimiyimana (sesn) <sesn@odoo.com>