[FIX] {sale_}account_taxcloud: do not recompute taxes on post for down payment
Steps to reproduce:
- Install Accounting and Sales
- Configure TaxCloud in Accounting settings
- Create a SO:
  * Customer: [US customer] (e.g. Azure Interior)
  * Product: [any]
  * Unit Price: [any] (e.g. $ 200.00)
  * Fiscal Position (Other Info tab): Automatic Tax Mapping (TaxCloud)
- Confirm the SO
=> The taxes will be recomputed with TaxCloud (e.g.  $ 20.50 with "Tax 10.250 %")
- Create a down payment from SO (via "Create Invoice" button):
  * Create Invoice: Down payment (percentage)
  * Down Payment Amount: 50%
=> The total amount of the down payment is 50% of the total amount of the SO (i.e. $ 110.25)
and the computed taxes are the same than the taxes on the SO.
- Confirm the down payment

Issue:
Upon confirmation, the taxes are recomputed with TaxCloud and the total amount is
not the same anymore.

Cause:
The "Down payment" product has a specific TaxCloud Category and as the taxes are
recomputed upon confirmation, the tax amount changes (e.g. from $ 20.50 to $ 0.00).

Solution:
Do not recompute taxes with TaxCloud when confirming a down payment as the taxes
should have already been computed correctly when the down payment has been created.

opw-3814524

closes odoo/enterprise#61178

X-original-commit: af381629b29e7a4c2da2c5544a52b2573a31b802
Signed-off-by: John Laterre (jol) <jol@odoo.com>
Signed-off-by: Anh Thao Pham (pta) <pta@odoo.com>